Added InverseDynamics block
One question arises:
The GetBiasForces block (and a useful GetGravityCompensation block) are simply calls to the inverse dynamics algorithm with appropriate parameters.
We could create only one C++ function (InverseDynamics).
One issue in this is how to define the zero constants for the joints (given that we do not know the size in Simulink but only in C++). We can multiply the Joint configuration input signal by a Zero-Gain. But this will of course affect computation (? maybe ?).
Another alternative is to create a dropdown parameter and let the user choose if he wants a full inverse dynamics, a bias forces or gravity compensation block. Only one C++, only one simulink block, but a little more complex logic.

cc @DanielePucci, @traversaro