Hopefully fixes https://github.com/robotology/WB-Toolbox/issues/15
mdlInitializeConditions is called also (strangely) in compilation.
Moved back the code to the outputs method.
Added additional option to ask to put the synchronizer block as soon as possible in the block order.
Note that this apply only INSIDE a subsystem. Controllers should take care that, if this block is in a subsystem of its own, it gets called asap (by using the priority option) @DanielePucci