Error class has been enhanced and renamed to Log
wbt::Log is the new class that provides log support to the WBToolbox.
It now supports both Errors and Warnings.

For being more flexible, wbt::Log is a singleton. This approach simplifies dramatically the handling of warnings, and can be called from every method without the need of passing it though arguments. Since singletons cannot be deleted, wbt::Log state is reset in the early stage of the processing.

[Squash] toolbox.cpp