New RobotInterface class
This class contains all the required data structures to operate on a robot, either through the model or the interfaces to the hardware / gazebo.

From the Simulink point of view, there is one instance of RobotInterface for every Configuration block added in the model. In fact, RobotInterface extends by composition the Configuration object.

It handles the pointers to its members in a lazy way, initializing on the fly the asked resource.