Retaining the RemoteControlBoardRemapper needs a unique port prefix
If there are two robot configurations in the same system which share some of the ControlBoards, the RemoteControlBoardRemapper devices cannot be allocated with the same port prefix.

As a workaround, a new function to generate an unique identifier for every Configuration object has been included. Under the hood it uses the Simulink's block name (which is unique).

Furthermore, the initialization of the Yarp Network was missing at this stage of the execution.