New logic for gathering RemoteControlBoardRemapper's devices and KinDynComputations
This commit moves the ownership of the RobotInterface object (created from wbt::Configuration) from
the ToolboxSingleton to the WBBlock which actually needs it. The ToolboxSingleton will hold only a
weak pointer to it, and it uses this weak pointer to share its ownership with other WBBlocks that share
the same Configuration (and hence the same RobotInterface object).

This new approach allows removing the former retain / release logic which was not very robust in edge cases
(e.g. when the robot crashes badly). In these cases, the entire toolbox reaches a dirty state because it
couldn't terminare properly, and restarting Matlab was required.

These modification prevent this status and greatly simplify the logic behind this configuration step. The cost
to pay is during the configureSizeAndPorts step, because now every block which needs a KinDynComputations object
needs to allocate it independently. After this step, however, the object is unique and shared. Considering that
at most this object is used to gather the number of DoFs, this shouldn't introduce any performance regression.