Updated blocks simplifying the retrieval of the RobotInterface object
The new logic for storing the RobotInterface assures that the object is stored as
a shared pointer directly into the WBBlock class. Before, instead, it was gathered
passing through a BlockInformation implementation.
Avoiding this step allows removing all the checks on the nullity of the RobotInterface object.

Furthermore, asking the KinDynComputations object now return a shared pointer, not anymore a weak pointer.