Renamed ToolboxSingleton to WholeBodySingleton
After the recent code refactorings, most of the singleton logic has been moved somewhere else,
mainly in the RobotInterface class. Right now it is only used for sharing among WBBlocks
the ownership of the RobotInterface object created from a wbt::Configuration.

Since it is something specific to the WBBlock class, it makes more sense call this singleton WholeBodySingleton.