Fixed memory leak
This leak was introduced after the switch to the plugin-based loading of the block library. In the mdlInitializeSizes step the block memory was not cleaned, and the associated RobotInterface block was never deallocated.

This causes a series of issues. Firstly, the toolbox never reaches a clean status when a simulation ends. Secondly, Matlab doesn't close properly and manually killing the process is necessary.