Fix sharedlibpp usage in S-Function
The factory object should not be deleted if a class is allocated from the loaded shared library. The reason is that the library is close and unloaded in the destructor of the factory object, and while the pointer to the interface remains valid (non null) it points to a memory area not owned by anyone.